+++ Assets/VRM/Runtime/IO/VRMExporter.cs
@@ -112,7 +112,7 @@ namespace VRM
                     VRM.meta.title = meta.Title;
                     if (meta.Thumbnail != null)
                     {
-                        VRM.meta.texture = TextureExporter.ExportTexture(glTF, glTF.buffers.Count - 1, meta.Thumbnail);
+                        VRM.meta.texture = glTF.PushGltfTexture(glTF.buffers.Count - 1, meta.Thumbnail);
                     }
 
                     VRM.meta.licenseType = meta.LicenseType;
